fix(arc-278): move react-query to service + fix file casing
