+++ public/locales/nl/common.json
@@ -59,15 +59,17 @@
   "pages/gebruiksvoorwaarden/index___annuleer": "Annuleer",
   "pages/gebruiksvoorwaarden/index___gebruiksvoorwaarden": "Gebruiksvoorwaarden",
   "pages/gebruiksvoorwaarden/index___gebruiksvoorwaarden-omschrijving": "Gebruiksvoorwaarden omschrijving",
+  "pages/index___geen-resultaten-voor-de-geselecteerde-filters": "Geen resultaten voor de geselecteerde filters.",
   "pages/index___hier-kom-je-er-alles-over-te-weten": "Hier kom je er alles over te weten.",
+  "pages/index___laden": "Laden...",
   "pages/index___plan-een-nieuw-bezoek-stap-fysiek-binnen-en-krijg-meteen-toegang-tot-het-digitale-archief-van-de-leeszaal-benieuwd-hoe-het-werkt": "Plan een nieuw bezoek, stap fysiek binnen en krijg meteen toegang tot het digitale archief van de leeszaal. Benieuwd hoe het werkt?",
   "pages/index___toon-alles-amount": "Toon alles ({{amount}})",
   "pages/index___vind-een-leeszaal": "Vind een leeszaal",
   "pages/index___welkom-in-de-digitale-leeszaal": "Welkom in de digitale leeszaal",
   "pages/index___zoek": "Zoek",
+  "pages/leeszaal/reading-room-slug/index___filters": "Filters",
   "pages/leeszaal/reading-room-slug/object-id/index___meer-info": "Meer info",
   "pages/leeszaal/reading-room-slug/object-id/index___sluit": "Sluit",
-  "pages/leeszaal/reading-room-slug/index___filters": "Filters",
   "pages/leeszaal/slug___wis-volledige-zoekopdracht": "Wis volledige zoekopdracht",
   "pages/leeszaal/slug___zoek-naar": "Zoek naar:",
   "pages/leeszaal/slug___zoek-op-trefwoord-jaartal-aanbieder": "Zoek op trefwoord, jaartal, aanbieder..."
