+++ src/modules/reading-room/services/reading-room/reading-room.service.ts
@@ -1,13 +1,26 @@
 import { stringify } from 'query-string';
 
+import { ApiResponseWrapper, ReadingRoomInfo } from '@reading-room/types';
 import { ApiService } from '@shared/services';
 
 import { READING_ROOM_SERVICE_BASE_URL } from './reading-room.service.const';
 
 class ReadingRoomService extends ApiService {
-	public async getAll(query?: string): Promise<unknown> {
-		const searchParams = query ? stringify({ query }) : {};
-		return await this.api.get('', { searchParams }).json();
+	public async getAll(
+		searchInput = '',
+		page = 0,
+		size = 20
+	): Promise<ApiResponseWrapper<ReadingRoomInfo>> {
+		const parsed = await this.api
+			.get(
+				`${READING_ROOM_SERVICE_BASE_URL}?${stringify({
+					query: `%${searchInput}%`,
+					page,
+					size,
+				})}`
+			)
+			.json();
+		return parsed as ApiResponseWrapper<ReadingRoomInfo>;
 	}
 
 	public async getById(roomId: string): Promise<unknown> {
